Modify README
README is updated with dependency freshness (how updated the dependencies are). It also has a section for Contributing and License (created in this commit).

remark tool is added to lint the README file (and future docs files).

package.json is updated with new test for remark as well as the existing eslint tool. Any errors reported from eslint is fixed in this commit.

Unnecessary commentin .gitattributes is removed.